refactor: colocate global app shell under app/_components
