+++ web/app/_components/body.tsx
@@ -1,10 +1,10 @@
 'use client'
 import { AppRouterCacheProvider } from '@mui/material-nextjs/v15-appRouter'
 import React, { Suspense } from 'react'
-import { DataProvider } from '../../lib/utils/data-context'
-import { MainContextProvider } from '../../lib/utils/main-context'
-import { UserContextProvider } from '../../lib/utils/user-context'
-import { ConfigProvider } from '../utils/config'
+import { ConfigProvider } from '@/lib/utils/config'
+import { DataProvider } from '@/lib/utils/data-context'
+import { MainContextProvider } from '@/lib/utils/main-context'
+import { UserContextProvider } from '@/lib/utils/user-context'
 import Main from './main'
 
 const Body = ({ children }: { children: React.ReactNode }) => {
