+++ web/app/_components/content-box.tsx
