+++ web/app/_components/main.tsx
@@ -10,14 +10,14 @@ import { createTheme, ThemeProvider } from '@mui/material/styles'
 import NextAdapterApp from 'next-query-params/app'
 import React, { useCallback, useEffect, useMemo, useRef, useState } from 'react'
 import { QueryParamProvider } from 'use-query-params'
-import NavBar from '@/lib/components/nav-bar'
-import { useConfig } from '../utils/config'
-import { useData } from '../utils/data-context'
-import { useMainContext } from '../utils/main-context'
-import { MapContextProvider } from '../utils/map-context'
-import { idToShowUrl } from '../utils/meta-utils'
+import { useConfig } from '@/lib/utils/config'
+import { useData } from '@/lib/utils/data-context'
+import { useMainContext } from '@/lib/utils/main-context'
+import { MapContextProvider } from '@/lib/utils/map-context'
+import { idToShowUrl } from '@/lib/utils/meta-utils'
 import BottomBar from './bottom-bar'
 import GMap from './map'
+import NavBar from './nav-bar'
 import ToolBox from './tool-box'
 
 const Main = ({ children }: { children: React.ReactNode }) => {
