+++ web/app/_components/map.test.tsx
@@ -3,25 +3,25 @@ import React from 'react'
 import '@testing-library/jest-dom'
 import { initialize } from '@googlemaps/jest-mocks'
 import { Mock } from 'vitest'
-import { useConfig } from '../utils/config'
-import { useData } from '../utils/data-context'
-import { useMainContext } from '../utils/main-context'
-import { useMapContext } from '../utils/map-context'
+import { useConfig } from '@/lib/utils/config'
+import { useData } from '@/lib/utils/data-context'
+import { useMainContext } from '@/lib/utils/main-context'
+import { useMapContext } from '@/lib/utils/map-context'
 import GMap from './map'
 
-vi.mock('../utils/main-context', () => ({
+vi.mock('@/lib/utils/main-context', () => ({
   useMainContext: vi.fn(),
 }))
 
-vi.mock('../utils/map-context', () => ({
+vi.mock('@/lib/utils/map-context', () => ({
   useMapContext: vi.fn(),
 }))
 
-vi.mock('../utils/config', () => ({
+vi.mock('@/lib/utils/config', () => ({
   useConfig: vi.fn(),
 }))
 
-vi.mock('../utils/data-context', () => ({
+vi.mock('@/lib/utils/data-context', () => ({
   useData: vi.fn(),
 }))
